+++ src/components/MainHeader.jsx
@@ -5,8 +5,6 @@ import { darkTheme, lightTheme } from '../theme';
 
 import SunIcon from '../icons/sun.svg';
 import MaskIcon from '../icons/mask.svg';
-import UserIcon from '../icons/UserIcon';
-import LoginIcon from '../icons/LoginIcon';
 
 import {
   BtnBox,
@@ -19,13 +17,65 @@ import {
   WeatherBtn,
   WeatherIcon,
   WeatherText
-} from '../styles/main-page/HeaderStyle';
+} from '../styles/mainPage/HeaderStyle';
+import axios from 'axios';
+
+const API_KEY = '947586767a6ce78304ecfd287c3de3ed';
 
 const MainHeader = () => {
   const dispatch = useDispatch();
   const location = useLocation();
   const isDarkMode = useSelector((state) => state.theme.isDarkMode);
+
   const [isScrolled, setIsScrolled] = useState(false);
+  // 미세먼지 상태관리
+  // const [airPollution, setAirPollution] = useState({});
+  const [showText, setShowText] = useState('');
+
+  // 현재위치데이터
+  const getUserLocation = () => {
+    const success = (position) => {
+      // console.log("위치정보: ", position);
+      const latitude = position.coords.latitude;
+      const longitude = position.coords.longitude;
+
+      getAirData(latitude, longitude);
+    };
+    const error = (error) => {
+      console.error('Error getting location:', error);
+    };
+    navigator.geolocation.getCurrentPosition(success, error);
+  };
+
+  // 미세먼지농도 데이터
+  const getAirData = async (latitude, longitude) => {
+    try {
+      const response = await axios.get(
+        `http://api.openweathermap.org/data/2.5/air_pollution?lat=${latitude}&lon=${longitude}&appid=${API_KEY}&units=metric&lang=kr`
+      );
+      // console.log(response.data.list[0].components); // pm10, pm2_5
+      const air = response.data.list[0].components;
+      // setAirPollution(air);
+      airCondition(air.pm2_5);
+    } catch (e) {
+      console.log(e.message);
+    }
+  };
+
+  // 대기질 표시
+  const airCondition = (pm2_5) => {
+    if (pm2_5 === undefined) {
+      setShowText('데이터 없음');
+    } else if (pm2_5 <= 15) {
+      setShowText('좋음');
+    } else if (pm2_5 <= 25) {
+      setShowText('보통');
+    } else if (pm2_5 <= 50) {
+      setShowText('나쁨');
+    } else {
+      setShowText('매우나쁨');
+    }
+  };
 
   const handleToggleWeather = () => {
     dispatch({ type: 'WEATHER_THEME' });
@@ -40,6 +90,9 @@ const MainHeader = () => {
   };
 
   useEffect(() => {
+    getUserLocation();
+    airCondition();
+
     window.addEventListener('scroll', handleScroll);
     return () => {
       window.removeEventListener('scroll', handleScroll);
@@ -70,12 +123,12 @@ const MainHeader = () => {
             {!isDarkMode ? (
               <BtnInner>
                 <WeatherIcon src={SunIcon} />
-                <WeatherText>미세먼지농도 좋음</WeatherText>
+                <WeatherText>현재 대기질은 {showText}, 뛰기 좋은 날입니다</WeatherText>
               </BtnInner>
             ) : (
               <BtnInner>
                 <WeatherIcon src={MaskIcon} />
-                <WeatherText>미세먼지농도 나쁨</WeatherText>
+                <WeatherText>현재 대기질은 {showText}, 외출을 자제해 주세요</WeatherText>
               </BtnInner>
             )}
           </WeatherBtn>
