--- src/components/main-page/Section02.jsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-
-const Section02 = () => {
-  return <Warpper>지도페이지-국봉</Warpper>;
-};
-
-export default Section02;
-
-// style
-const Warpper = styled.div`
-  width: 100%;
-  height: 100vh;
-  background-color: #fff;
-  border-top: 3px solid red;
-`;
