+++ src/components/mainPage/Section01.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import FilterMenu from './section01/FilterMenu';
+import MeetingList from './section01/MeetingList';
+
+import { Warpper } from '../../styles/mainPage/MainStyle';
+
+const Section01 = () => {
+  return (
+    <Warpper>
+      <FilterMenu />
+      <MeetingList />
+    </Warpper>
+  );
+};
+
+export default Section01;
