+++ src/pages/Home.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import Banner from '../components/main-page/Banner';
-import Section01 from '../components/main-page/Section01';
-import Section02 from '../components/main-page/Section02';
+import Banner from '../components/mainPage/Banner';
+import Section01 from '../components/mainPage/Section01';
+import Section02 from '../components/mainPage/Section02';
 
 const Home = () => {
   return (
