+++ src/styles/mainPage/MainStyle.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const StyledIcon = styled.svg`
+  width: 1.5rem;
+  height: 1.5rem;
+  stroke: ${({ theme }) => theme.textColor};
+`;
+export const Warpper = styled.div`
+  width: 100%;
+  background-color: #fff;
+`;
